requirements.txt
phidata==2.7.7
python-dotenv
duckduckgo-search
groq
openai
streamlit
alpha-vantage==3.0.0
requests